Move public page response interface to lib/pages
